refactor: extract Telegram config repair logic to separate module
- Extract Telegram-related config validation and repair functions
  from doctor-config-flow.ts (2106 lines) to doctor-config-telegram.ts (253 lines)
- Reduces doctor-config-flow.ts to 1865 lines
- Improves code organization and maintainability
- All tests pass
